requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.0.3
Werkzeug==3.1.3
psycopg[binary]==3.1.19
requests==2.32.4
python-dotenv==1.0.1

